Scraping for VA
The first file scrapes data based on year, uses the base code from the original given to us.

The other I changed more significantly to take a date argument in string form or date form in the format "yyyy-mm-dd" to both map and calculate based on data from three days before up to the day in question. The plan is to be able to feed the current date from python into the R file, invoke the file to make the calculations (maybe we don't need the mapping) and return the outputted data to add into the dataframe.

Question though, should I make it only consider the one day rather than multiple, or should it find the minimum across those days, maximum across those days, and average of all those days?